+++ x-pack/plugins/observability_solution/observability_onboarding/e2e/cypress/e2e/logs/custom_logs/configure.cy.ts
@@ -5,8 +5,7 @@
  * 2.0.
  */
 
-// Failing: See https://github.com/elastic/kibana/issues/183341
-describe.skip('[Logs onboarding] Custom logs - configure step', () => {
+describe('[Logs onboarding] Custom logs - configure step', () => {
   describe('logFilePaths', () => {
     beforeEach(() => {
       cy.loginAsViewerUser();
@@ -20,7 +19,7 @@ describe.skip('[Logs onboarding] Custom logs - configure step', () => {
       });
 
       it('navigates to observability logs onboarding page', () => {
-        cy.getByTestSubj('observabilityOnboardingBackButtonBackButton').click();
+        cy.getByTestSubj('observabilityOnboardingFlowBackToSelectionButton').click();
 
         cy.url().should('include', '/app/observabilityOnboarding');
       });
