+++ x-pack/plugins/observability_solution/observability_onboarding/public/application/observability_onboarding_flow.tsx
@@ -5,12 +5,11 @@
  * 2.0.
  */
 
-import { i18n } from '@kbn/i18n';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import React, { useEffect } from 'react';
 import { Route, Routes } from '@kbn/shared-ux-router';
-import { useNavigate, useLocation } from 'react-router-dom-v5-compat';
-import { EuiButtonEmpty, EuiPageTemplate, EuiSpacer } from '@elastic/eui';
+import { useLocation } from 'react-router-dom-v5-compat';
+import { EuiPageTemplate, EuiSpacer } from '@elastic/eui';
 import { css } from '@emotion/react';
 import backgroundImageUrl from './header/background.svg';
 import { Footer } from './footer/footer';
@@ -18,6 +17,7 @@ import { OnboardingFlowForm } from './onboarding_flow_form/onboarding_flow_form'
 import { Header } from './header/header';
 import { SystemLogsPanel } from './quickstart_flows/system_logs';
 import { CustomLogsPanel } from './quickstart_flows/custom_logs';
+import { BackButton } from './shared/back_button';
 
 const queryClient = new QueryClient();
 
@@ -68,25 +68,3 @@ export function ObservabilityOnboardingFlow() {
     </QueryClientProvider>
   );
 }
-
-const BackButton = () => {
-  const navigate = useNavigate();
-  const location = useLocation();
-
-  return (
-    <>
-      <EuiButtonEmpty
-        data-test-subj="observabilityOnboardingExperimentalOnboardingFlowBackToSelectionButton"
-        iconType="arrowLeft"
-        flush="left"
-        onClick={() => navigate(`../${location.search}`)}
-      >
-        {i18n.translate(
-          'xpack.observability_onboarding.experimentalOnboardingFlow.button.backToSelectionLabel',
-          { defaultMessage: 'Back to selection' }
-        )}
-      </EuiButtonEmpty>
-      <EuiSpacer size="m" />
-    </>
-  );
-};
