+++ x-pack/plugins/observability_solution/observability_onboarding/public/application/quickstart_flows/custom_logs/inspect.tsx
@@ -11,16 +11,16 @@ import { i18n } from '@kbn/i18n';
 import { EuiTitle, EuiSpacer } from '@elastic/eui';
 import { StepModal } from '../shared/step_panel';
 import { useWizard } from '.';
-import { BackButton } from '../shared/back_button';
+import { BackButton } from '../../shared/back_button';
 
 export function Inspect() {
-  const { goBack, getState, getPath, getUsage } = useWizard();
+  const { getState, getPath, getUsage } = useWizard();
   return (
     <StepModal
       title={i18n.translate('xpack.observability_onboarding.inspect.stepPanel.inspectWizardLabel', {
         defaultMessage: 'Inspect wizard',
       })}
-      panelFooter={[<BackButton onBack={goBack} />]}
+      panelFooter={[<BackButton />]}
     >
       <EuiTitle size="s">
         <h3>
