+++ x-pack/plugins/observability_solution/observability_onboarding/public/application/shared/back_button.tsx
@@ -1,0 +1,33 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { i18n } from '@kbn/i18n';
+import React from 'react';
+import { useNavigate, useLocation } from 'react-router-dom-v5-compat';
+import { EuiButtonEmpty, EuiSpacer } from '@elastic/eui';
+
+export const BackButton = () => {
+  const navigate = useNavigate();
+  const location = useLocation();
+
+  return (
+    <>
+      <EuiButtonEmpty
+        data-test-subj="observabilityOnboardingFlowBackToSelectionButton"
+        iconType="arrowLeft"
+        flush="left"
+        onClick={() => navigate(`../${location.search}`)}
+      >
+        {i18n.translate(
+          'xpack.observability_onboarding.experimentalOnboardingFlow.button.backToSelectionLabel',
+          { defaultMessage: 'Back to selection' }
+        )}
+      </EuiButtonEmpty>
+      <EuiSpacer size="m" />
+    </>
+  );
+};
